Develop pamela

+++ Models/Reserva.cs
@@ -15,13 +15,13 @@ namespace ProjetoHotelSerranoSenac.Models
         public Quarto Quarto { get; set; }
         public DateTime DataCheckin { get; set; }
         public DateTime DataCheckout { get; set; }
-        public decimal Preco { get; set; }
+        public double Preco { get; set; }
         public int HotelId { get; set; }
         public Hotel Hotel { get; set; }
 
         public Reserva() { }
 
-        public Reserva(int clienteId, int quartoId, DateTime dataCheckin, DateTime dataCheckout, decimal preco, int hotelId)
+        public Reserva(int clienteId, int quartoId, DateTime dataCheckin, DateTime dataCheckout, double preco, int hotelId)
         {
             ClienteId = clienteId;
             QuartoId = quartoId;
